robot/requirements.txt
Pillow==10.0.0
python-ev3dev2==2.1.0.post1
